Move distribution to daemon/internal/distribution
