+++ daemon/containerd/image_pull.go
@@ -16,8 +16,8 @@ import (
 	"github.com/containerd/log"
 	"github.com/containerd/platforms"
 	"github.com/distribution/reference"
+	"github.com/docker/docker/daemon/internal/distribution"
 	"github.com/docker/docker/daemon/internal/metrics"
-	"github.com/docker/docker/distribution"
 	"github.com/docker/docker/errdefs"
 	"github.com/docker/docker/pkg/progress"
 	"github.com/docker/docker/pkg/streamformatter"
