+++ daemon/containerd/service.go
@@ -17,8 +17,8 @@ import (
 	"github.com/docker/docker/daemon/container"
 	daemonevents "github.com/docker/docker/daemon/events"
 	dimages "github.com/docker/docker/daemon/images"
+	"github.com/docker/docker/daemon/internal/distribution"
 	"github.com/docker/docker/daemon/snapshotter"
-	"github.com/docker/docker/distribution"
 	"github.com/docker/docker/errdefs"
 	"github.com/moby/sys/user"
 	"github.com/opencontainers/go-digest"
