+++ daemon/images/image_push.go
@@ -7,9 +7,9 @@ import (
 
 	"github.com/distribution/reference"
 	"github.com/docker/distribution/manifest/schema2"
+	"github.com/docker/docker/daemon/internal/distribution"
+	progressutils "github.com/docker/docker/daemon/internal/distribution/utils"
 	"github.com/docker/docker/daemon/internal/metrics"
-	"github.com/docker/docker/distribution"
-	progressutils "github.com/docker/docker/distribution/utils"
 	"github.com/docker/docker/pkg/progress"
 	"github.com/moby/moby/api/types/backend"
 	"github.com/moby/moby/api/types/registry"
