+++ daemon/images/service.go
@@ -11,10 +11,10 @@ import (
 	"github.com/containerd/log"
 	"github.com/docker/docker/daemon/container"
 	daemonevents "github.com/docker/docker/daemon/events"
+	"github.com/docker/docker/daemon/internal/distribution"
+	"github.com/docker/docker/daemon/internal/distribution/metadata"
+	"github.com/docker/docker/daemon/internal/distribution/xfer"
 	"github.com/docker/docker/daemon/internal/layer"
-	"github.com/docker/docker/distribution"
-	"github.com/docker/docker/distribution/metadata"
-	"github.com/docker/docker/distribution/xfer"
 	"github.com/docker/docker/image"
 	refstore "github.com/docker/docker/reference"
 	"github.com/opencontainers/go-digest"
