+++ daemon/internal/builder-next/adapters/containerimage/pull.go
@@ -26,9 +26,9 @@ import (
 	"github.com/containerd/platforms"
 	"github.com/distribution/reference"
 	dimages "github.com/docker/docker/daemon/images"
+	"github.com/docker/docker/daemon/internal/distribution/metadata"
+	"github.com/docker/docker/daemon/internal/distribution/xfer"
 	"github.com/docker/docker/daemon/internal/layer"
-	"github.com/docker/docker/distribution/metadata"
-	"github.com/docker/docker/distribution/xfer"
 	"github.com/docker/docker/image"
 	pkgprogress "github.com/docker/docker/pkg/progress"
 	refstore "github.com/docker/docker/reference"
