+++ daemon/internal/builder-next/worker/worker.go
@@ -17,9 +17,9 @@ import (
 	imageadapter "github.com/docker/docker/daemon/internal/builder-next/adapters/containerimage"
 	mobyexporter "github.com/docker/docker/daemon/internal/builder-next/exporter"
 	"github.com/docker/docker/daemon/internal/builder-next/worker/mod"
+	distmetadata "github.com/docker/docker/daemon/internal/distribution/metadata"
+	"github.com/docker/docker/daemon/internal/distribution/xfer"
 	"github.com/docker/docker/daemon/internal/layer"
-	distmetadata "github.com/docker/docker/distribution/metadata"
-	"github.com/docker/docker/distribution/xfer"
 	pkgprogress "github.com/docker/docker/pkg/progress"
 	"github.com/moby/buildkit/cache"
 	cacheconfig "github.com/moby/buildkit/cache/config"
