+++ daemon/internal/distribution/errors.go
@@ -14,7 +14,7 @@ import (
 	v2 "github.com/docker/distribution/registry/api/v2"
 	"github.com/docker/distribution/registry/client"
 	"github.com/docker/distribution/registry/client/auth"
-	"github.com/docker/docker/distribution/xfer"
+	"github.com/docker/docker/daemon/internal/distribution/xfer"
 	"github.com/docker/docker/errdefs"
 	"github.com/pkg/errors"
 )
