+++ daemon/internal/distribution/errors_test.go
