+++ daemon/internal/distribution/manifest.go
