+++ daemon/internal/distribution/manifest_test.go
