+++ daemon/internal/distribution/metadata/metadata.go
