+++ daemon/internal/distribution/metadata/v2_metadata_service.go
