+++ daemon/internal/distribution/pull_v2_windows.go
