+++ daemon/internal/distribution/push_v2_test.go
@@ -10,8 +10,8 @@ import (
 	"github.com/docker/distribution"
 	"github.com/docker/distribution/manifest/schema2"
 	"github.com/docker/distribution/registry/api/errcode"
+	"github.com/docker/docker/daemon/internal/distribution/metadata"
 	"github.com/docker/docker/daemon/internal/layer"
-	"github.com/docker/docker/distribution/metadata"
 	"github.com/docker/docker/pkg/progress"
 	refstore "github.com/docker/docker/reference"
 	registrypkg "github.com/docker/docker/registry"
