+++ daemon/internal/distribution/registry.go
