+++ daemon/internal/distribution/registry_unit_test.go
