+++ daemon/internal/distribution/utils/progress.go
