+++ daemon/internal/distribution/xfer/download.go
