+++ daemon/pkg/plugin/fetch_linux.go
@@ -13,7 +13,7 @@ import (
 	cerrdefs "github.com/containerd/errdefs"
 	"github.com/containerd/log"
 	"github.com/distribution/reference"
-	progressutils "github.com/docker/docker/distribution/utils"
+	progressutils "github.com/docker/docker/daemon/internal/distribution/utils"
 	"github.com/docker/docker/pkg/ioutils"
 	"github.com/docker/docker/pkg/progress"
 	"github.com/docker/docker/pkg/stringid"
