+++ daemon/server/router/distribution/distribution_routes.go
@@ -9,8 +9,8 @@ import (
 	"github.com/docker/distribution"
 	"github.com/docker/distribution/manifest/manifestlist"
 	"github.com/docker/distribution/manifest/schema2"
+	distributionpkg "github.com/docker/docker/daemon/internal/distribution"
 	"github.com/docker/docker/daemon/server/httputils"
-	distributionpkg "github.com/docker/docker/distribution"
 	"github.com/docker/docker/errdefs"
 	"github.com/moby/moby/api/types/registry"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
